hardware/mordomo - (orange-pi-5-16gb)/ecossistemas/mordomo/containers/core/mordomo-orchestrator/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
nats-py==2.7.0
pydantic==2.6.0
pydantic-settings==2.1.0
litellm==1.18.0
sentence-transformers==2.3.1
faiss-cpu==1.7.4
numpy==1.26.3
python-dotenv==1.0.0
httpx==0.26.0
asyncpg==0.29.0
sqlalchemy==2.0.25
python-consul==1.1.0
